refactor: 컴포넌트 파일명 카멜 케이스 변경, 카카오 로그인 호버 로고 색상, 랜딩 페이지 발자국, 로고 svg 파일로 변경
