+++ src/app/globals.css
@@ -45,7 +45,6 @@
       var(--color-white-50),
       var(--color-white-0)
     );
-    opacity: 0.8;
     font-family: "Pretendard", sans-serif;
   }
 }
