+++ src/components/ui/paw/paw.tsx
@@ -1,5 +1,5 @@
 "use client";
-import PawPrints from "./paw-prints";
+import PawPrints from "./PawPrints";
 export default function Paw() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-center p-24 ">
